Add more real(kind=DP), mostly to do with concentration values.

+++ configFunctions.f90
@@ -115,7 +115,7 @@ SUBROUTINE setConcentrations (refSpeciesNames, concSpeciesNames, &
   CHARACTER(LEN=maxSpecLength), intent(in) :: concSpeciesNames(:), refSpeciesNames(:)
   CHARACTER(LEN=maxSpecLength) :: k, m
   real(kind=DP), intent(in) :: inputConcentrations(:)
-  DOUBLE PRECISION, intent(out) :: outputConcentrations(:)
+  real(kind=DP), intent(out) :: outputConcentrations(:)
   INTEGER(kind=NPI) :: j, i
   LOGICAL :: match
